requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
ffmpy==0.3.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
moviepy==1.0.3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
